requirements.txt
asgiref==3.5.2
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
mysqlclient==2.1.1
Pillow==9.2.0
PyMySQL==1.0.2
python-decouple==3.5
python-dotenv==0.19.2
pytz==2022.2.1
sqlparse==0.4.2
whitenoise==6.2.0
